0923张松光第3次上传

+++ api-demo/src/main/java/com/jk/model/Book.java
@@ -46,7 +46,7 @@ public class Book implements Serializable {
 
     private Integer bookCount;
 
-    private Double hits;
+    private Integer hits;
 
     private Integer bookStar;
